requirements.txt
google-api-python-client==2.100.0
google-auth==2.25.2
google-auth-oauthlib==1.1.0
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.24.4
schedule==1.2.0
tqdm==4.66.1
pyyaml==6.0.1
requests==2.31.0
